Fixed order that composer install is run and vendor files are renamed
I was renaming vendor to vendor-dev and running composer install
then right after that I was renaming vendor-dev back to vendor
and then copying vendor files to build folder which is incorrect.

Changed order of these tasks so it's:
- rename vendor to vendor-dev
- run composer install
- copy files to build
- rename vendor-dev to vendor

Also doing a better job at handling errors when vendor* files can't be
copied.