Hooks and setup for Authors feature
Added all types of hooks to save, load recipe with author info, save
author info into settings, etc.

One major kind of change is the way database upgrade is done. Since
dbDelta doesn't actually delete columns, I removed the guard to see
which table version is installed and only do changes if there's a version
change. The reason for this is because plugins, such as Authors plugin
needs to be able to modify database schema and tracking the version
that the plugin wants + the version that Zip Recipes core wants is tricky
and not worth doing.

Also, removed 'init' hook since it was being called at wrong time for
Author plugin to work. It's not needed anyway, since the `upgraer_process_complete`
hook works.