Registration and shortcode fix (#55)
This was done by Alex, I'm just squashing these commits into one since it's kind of messy:

Registration is now optional and can be skipped. It shows up when user creates/edits a recipe as well as in settings. Optional in both.

Also, it inits a shortcode so [zip-recipes ..] can be used by different plugins.