Removed amdZLRecipeInsertIntoPostEditor
It looks like it's not really needed. I thought it's used to create
the Zip Recipes placeholder inside the Visual mode of the WP Editor but
that's no longer the case.
The placeholder is inserted by
`_convert_codes_to_imgs` function which is attached to tinymce as
`tinymce.plugins.zrdnEditRecipe`.